+++ dao/projectDao.js
@@ -34,7 +34,7 @@ function _getProjectList (filterObj = {}) {
   let offset = filterObj.offset ? filterObj.offset : 0;
   let limit = filterObj.limit ? filterObj.limit : null;
   let sortby = filterObj.sortby ? filterObj.sortby : 'projectId';
-  let order = filterObj.order ? filterObj.order : true;
+  let order = filterObj.order !== null ? filterObj.order : true;
   let detail = filterObj.detail ? filterObj.detail : false;
 
   let sql = squel.select().from('scada.project_list', 'Project');
@@ -105,7 +105,7 @@ function _checkProjectRightByUserName (userName, filterObj = {}) {
   sql.left_join(squel.select().from('scada.user_allow_device', 'UserAllowDevice')
   .field('UserAllowDevice.user_id', 'userId').field('UserAllowDevice.proj_id', 'projectId')
   .join('scada.user_info', 'UserInfo', squel.expr().and('UserAllowDevice.user_id = UserInfo.user_id').and('Userinfo.user_name = ?', userName)), 'SUB', 'SUB.projectId = Project.proj_id');
-  if (filterObj.projectId) {
+  /* if (filterObj.projectId) {
     sql.where(('UserAllowDevice.proj_id = ?', filterObj.projectId));
   }
   if (filterObj.scadaId) {
@@ -113,7 +113,7 @@ function _checkProjectRightByUserName (userName, filterObj = {}) {
   }
   if (filterObj.deviceId) {
     sql.where(('UserAllowDevice.device_id = ?', filterObj.deviceId));
-  }
+  } */
 
   return new Promise((resolve, reject) => {
     _sequelize.query(sql.toString(), { type: _sequelize.QueryTypes.SELECT, model: userAllowDeviceVo }).then((raws) => {